+++ dsf-bpe/dsf-bpe-server/src/main/java/org/highmed/dsf/fhir/questionnaire/EventQuestionnaireResponseHandler.java
@@ -1,6 +1,6 @@
 package org.highmed.dsf.fhir.questionnaire;
 
-import org.highmed.dsf.fhir.websocket.EventResourceHandler;
+import org.highmed.dsf.fhir.subscription.EventResourceHandler;
 import org.highmed.dsf.fhir.websocket.LastEventTimeIo;
 import org.highmed.dsf.fhir.websocket.ResourceHandler;
 import org.hl7.fhir.r4.model.DomainResource;
