+++ dsf-bpe/dsf-bpe-server/src/main/java/org/highmed/dsf/fhir/subscription/EventResourceHandler.java
@@ -1,4 +1,4 @@
-package org.highmed.dsf.fhir.websocket;
+package org.highmed.dsf.fhir.subscription;
 
 import org.hl7.fhir.r4.model.DomainResource;
